making a fetch button but... its giving error....

+++ client/src/popup/script.ts
@@ -1,6 +1,7 @@
 const website_notifyBtn = document.getElementById("website-notifyBtn")!;
 const api_notifyBtn = document.getElementById("api-notifyBtn")!;
 const discordBot_notifyBtn = document.getElementById("discord-bot-notifyBtn")!;
+const fetchBtn = document.getElementById("fetchBtn")!;
 
 website_notifyBtn.addEventListener("click", () => {
     if (website_notifyBtn.innerText === "Enable") {
@@ -34,21 +35,26 @@ discordBot_notifyBtn.addEventListener("click", () => {
     )!;
 });
 
-fetch(`${process.env.SERVER_URL}/notifications`, { method: "GET" })
-    .then((response) => {
-        return response.json();
-    })
-    .then((data) => {
-        data?.map((release: any, index: number) => {
-            const name: string = release.repository.name
-                .toLowerCase()
-                .split("jak-")[0];
+fetchBtn.addEventListener("click", () => {
+    fetch(`${process.env.SERVER_URL}/notifications`, { method: "GET" })
+        .then((response) => {
+            return response.json();
+        })
+        .then((data) => {
+            if (data)
+                data.map((release: any, index: number) => {
+                    const name: string = release.repository.name
+                        .toLowerCase()
+                        .split("jak-")[0];
 
-            if (localStorage.getItem(`${name}-notifyBtn`) === "Enable") {
-                chrome.runtime.sendMessage({ release, name });
-            }
-            fetch(`${process.env.SERVER_URL}/notifications/${index}`, {
-                method: "DELETE",
-            });
+                    if (
+                        localStorage.getItem(`${name}-notifyBtn`) === "Enable"
+                    ) {
+                        chrome.runtime.sendMessage({ release, name });
+                    }
+                    fetch(`${process.env.SERVER_URL}/notifications/${index}`, {
+                        method: "DELETE",
+                    });
+                });
         });
-    });
+});
